tests/wpt/tests/css/css-sizing/stretch/indefinite-3.html

<!DOCTYPE html>
<link rel="help"
  href="https://drafts.csswg.org/css-sizing-4/#stretch-fit-sizing">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/11006">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<meta name="assert"
  content="with indefinite available space, max-height: stretch resolves to infinity">

<body
  onload="checkLayout('[data-expected-client-height]')">

  <div style="max-height: stretch; height: 4000px; width: 200px; border: solid;"
    data-expected-client-height="4000">
    lorem ipsum</div>
